move animation to another component
