+++ web/src/components/LeagueLogo.vue
@@ -1,9 +1,8 @@
 <template>
-  <img :src="leagueLogo" :alt="league.name" :title="league.name" />
+  <img :src="leagueLogo" v-on:click="translate($event)" :alt="league.name" :title="league.name" />
 </template>
 
 <script>
-
 function getLogo(league) {
 	return `/static/img/leaguesLogos/${league}.png`
 }
@@ -17,6 +16,12 @@ export default {
   		leagueLogo: getLogo(this.league.slug)
   	}
   },
+
+  methods: {
+  	translate(event) {
+  		this.$emit('translate', event.target);
+  	}
+  }
 };
 </script>
